DB-3606: CR revisions
