--- packages/nextjs-kit/src/__tests__/snapshotTests/test.test.tsx
@@ -1,15 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Test from '../../components/test';
-
-/**
- * @vitest-environment jsdom
- */
-
-describe('<Test />', () => {
-  it("should render 'test'", () => {
-    const { asFragment } = render(<Test greeting="hello" title="world" />);
-
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
