--- packages/nextjs-kit/src/lib/interfaces.ts
@@ -1,10 +1,0 @@
-export interface Headers {
-  title: string;
-  href: string;
-}
-
-export interface Menu {
-  label: string;
-  path: string;
-  id: string;
-}
