+++ packages/nextjs-kit/tailwind.config.js
@@ -1,8 +1,5 @@
 module.exports = {
-  content: [
-    './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',
-  ],
+  content: ['./components/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {},
   },
